+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to the package will be documented in this file.
 
 ## [v2.7.1](https://github.com/upmind-automation/provision-provider-domain-names/releases/tag/v2.7.1) - 2023-07-03
 
+- Fix RealtimeRegister provider identifier
 - Disable RealtimeRegister `poll()`
+- Add RealtimeRegister logo, update description
 
 ## [v2.7.0](https://github.com/upmind-automation/provision-provider-domain-names/releases/tag/v2.7.0) - 2023-07-03